% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CInLPN2.default.R
\name{CInLPN2.default}
\alias{CInLPN2.default}
\title{Function that start estimation and prediction tasks}
\usage{
CInLPN2.default(
  fixed_X0.models,
  fixed_DeltaX.models,
  randoms_X0.models,
  randoms_DeltaX.models,
  mod_trans.model,
  DeltaT,
  outcomes,
  nD,
  mapping.to.LP,
  link,
  knots = NULL,
  subject,
  data,
  Time,
  Survdata = NULL,
  basehaz = NULL,
  knots_surv = NULL,
  assoc = 0,
  truncation = FALSE,
  fixed.survival.models = NULL,
  makepred,
  MCnr,
  type_int = NULL,
  sequence = NULL,
  ind_seq_i = NULL,
  nmes = NULL,
  cholesky = FALSE,
  paras.ini = NULL,
  indexparaFixeUser,
  paraFixeUser,
  maxiter,
  zitr,
  ide,
  univarmaxiter,
  nproc = 1,
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 0.001,
  print.info = FALSE,
  TimeDiscretization = TRUE,
  Tentry = NULL,
  Event = NULL,
  StatusEvent = NULL,
  ...
)
}
\arguments{
\item{fixed_X0.models}{fixed effects in the submodel for the baseline level of processes}

\item{fixed_DeltaX.models}{a two-sided linear formula object for specifying the response outcomes (one the left part of ~ symbol) 
and the covariates with fixed-effects (on the right part of ~ symbol) 
in the submodel for change over time of latent processes}

\item{randoms_X0.models}{random effects in the submodel for the baseline level of processes}

\item{randoms_DeltaX.models}{random effects in the submodel for change over time of latent processes}

\item{mod_trans.model}{model for elements of the temporal transition matrix, which captures 
the temporal influences between latent processes}

\item{DeltaT}{indicates the discretization step}

\item{outcomes}{indicates names of the outcomes}

\item{nD}{number of the latent processes}

\item{mapping.to.LP}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{link}{indicates link used to transform outcome}

\item{knots}{indicates position of knots used to transform outcomes}

\item{subject}{indicates the name of the covariate representing the grouping structure}

\item{data}{indicates the data frame containing all the variables for estimating the model}

\item{Time}{indicates the name of the covariate representing the time}

\item{Survdata}{Dataset for event and statusEvent}

\item{basehaz}{baseline hazard type (Weibull or splines)}

\item{knots_surv}{knots for splines (baseline hazard function)}

\item{assoc}{association type between outcomes and times-to-events}

\item{truncation}{boolean indicating if left truncation or not}

\item{fixed.survival.models}{fixed effects in the submodel for the event hazards}

\item{makepred}{indicates if predictions in the real scales of outcomes have to be done}

\item{MCnr}{number of replicates  to compute the predictions in the real scales of the outcomes}

\item{type_int}{type of Monte Carlo integration method to use}

\item{paras.ini}{initial values for parameters, default values is NULL}

\item{indexparaFixeUser}{position of parameters to be constrained}

\item{paraFixeUser}{values associated to the index of parameters to be constrained}

\item{maxiter}{maximum iteration}

\item{univarmaxiter}{maximum iteration for estimating univariate model}

\item{nproc}{number of processor to be used for running this package, default value is 1}

\item{epsa}{threshold for the convergence criterion on the parameters, default value is 1.e-4}

\item{epsb}{threshold for the convergence criterion on the likelihood, default value is 1.e-4}

\item{epsd}{threshold for the convergence criterion on the derivatives, default value is 1.e-3}

\item{print.info}{to print information during the liklihood optimization, default value is FALSE}

\item{\dots}{optional parameters}
}
\value{
CInLPN2 object
}
\description{
Function that start estimation and prediction tasks
}
