% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynNet-package.R
\docType{package}
\name{DynNet-package}
\alias{DynNet-package}
\title{A package for Causal Inference in a Latent Processes Network}
\description{
This package Contains functions to fit a dynamic model for multiple latent processes. 
The methodology is based on the combinaison 
of a linear mixed model for the latent processes trajectories and a system of 
differences equations for assessing their temporal influences. 
The estimation is done in the maximum likelihood framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anarouanet/DynNet}
}

}
\keyword{"Causality",}
\keyword{"Dynamic}
\keyword{Latent}
\keyword{data"}
\keyword{longitudinal}
\keyword{model","}
\keyword{multivariate}
\keyword{processes","}
