% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parametre.R
\name{f_paras.ini}
\alias{f_paras.ini}
\title{Initialisation of parameters}
\usage{
f_paras.ini(
  data,
  outcomes,
  mapped.to.LP,
  fixed_X0.models,
  fixed_DeltaX.models,
  randoms_DeltaX.models,
  randoms_X0.models,
  nb_RE,
  mod_trans.model,
  subject,
  MCnr = NULL,
  type_int = NULL,
  Survdata = NULL,
  basehaz = NULL,
  Time,
  link,
  knots,
  zitr = NULL,
  ide = NULL,
  DeltaT,
  maxiter = 25,
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 1e-04,
  nproc = 1,
  print.info = TRUE,
  TimeDiscretization = FALSE,
  fixed.survival.models = NULL,
  Tentry = NULL,
  Event = NULL,
  StatusEvent = NULL,
  assocT = NULL,
  truncation = NULL
)
}
\arguments{
\item{data}{indicates the data frame containing all the variables for estimating the model}

\item{outcomes}{names of the outcomes}

\item{mapped.to.LP}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{fixed_X0.models}{fixed effects in the submodel for the baseline level of processes}

\item{fixed_DeltaX.models}{a two-sided linear formula object for specifying the response outcomes (one the left part of ~ symbol) 
and the covariates with fixed-effects (on the right part of ~ symbol)}

\item{randoms_DeltaX.models}{random effects in the submodel for change over time of latent processes}

\item{randoms_X0.models}{random effects in the submodel for the baseline level of processes}

\item{nb_RE}{number of random effects}

\item{mod_trans.model}{model for elements of the temporal transition matrix, which captures 
the temporal influences between latent processes}

\item{subject}{indicates the name of the covariate representing the grouping structure}

\item{MCnr}{number Quasi-Monte Carlo replicates for the integration over random effects}

\item{type_int}{type of Monte Carlo integration method to use}

\item{Survdata}{database for survival sub-model}

\item{basehaz}{type of baseline hazard}

\item{Time}{indicates the name of the covariate representing the time}

\item{link}{indicates link used to transform outcome}

\item{knots}{indicates position of knots used to transform outcomes}

\item{zitr}{min and max of ordinal outcomes}

\item{ide}{vector of observed values for ordinal outcomes}

\item{DeltaT}{indicates the discretization step}

\item{maxiter}{maximum iterations}

\item{epsa}{threshold for the convergence criterion on the parameters, default value is 1.e-4}

\item{epsb}{threshold for the convergence criterion on the likelihood, default value is 1.e-4}

\item{epsd}{threshold for the convergence criterion on the derivatives, default value is 1.e-3}

\item{nproc}{number of processor to be used for running this package}

\item{print.info}{to print information during the liklihood optimization, default value is FALSE}

\item{TimeDiscretization}{a boolean indicating if the initial time has to be discretized (TRUE by default). When setting to FALSE, it allows to avoid discretization when running univariate model during parameter initialization.}

\item{fixed.survival.models}{specification of survival models}

\item{Tentry}{name of the variable of entry time}

\item{Event}{name of the variable of event time}

\item{StatusEvent}{name of the variable of event status}

\item{assocT}{specification of association between longitudinal and survival models}

\item{truncation}{boolean for delayed entry}
}
\value{
a list
}
\description{
Initialisation of parameters
}
