% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Loglik}
\alias{Loglik}
\title{Function that computes the log-likelihood of the observed data}
\usage{
Loglik(
  K,
  nD,
  mapping,
  paraOpt,
  paraFixe,
  posfix,
  paras_k,
  sequence,
  type_int,
  ind_seq_i,
  MCnr,
  nmes,
  m_is,
  Mod_MatrixY,
  Mod_MatrixYprim,
  df,
  x,
  z,
  q,
  nb_paraD,
  x0,
  z0,
  q0,
  cholesky,
  data_surv,
  data_surv_intY,
  nYsurv,
  basehaz,
  knots_surv,
  np_surv,
  survival,
  assoc,
  truncation,
  nE,
  Xsurv1,
  Xsurv2,
  if_link,
  zitr,
  ide,
  tau,
  tau_is,
  modA_mat,
  DeltaT,
  ii
)
}
\arguments{
\item{K}{an integer indicating the number of markers}

\item{nD}{an integer indicating the number of latent processes}

\item{mapping}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{paraOpt}{initial values for model parameters}

\item{paraFixe}{values associated to the index of parameters to be constrained}

\item{posfix}{position of parameters to be constrained}

\item{m_is}{vector of numbers of visit occasions for individuals}

\item{Mod_MatrixY}{model.matrix from markers transformation submodels}

\item{Mod_MatrixYprim}{model.matrix from the derivates of markers transformation submodels}

\item{df}{vector of numbers of parameters for each transformation model}

\item{x}{model.matrix for change's fixed submodel}

\item{z}{model.matrix for change's random effects submodel}

\item{q}{a vector of number of random effects on each change latent process over time}

\item{nb_paraD}{number of paramerters of the variance-covariance matrix of random effects}

\item{x0}{model.matrix for baseline's fixed submodel}

\item{z0}{model.matrix for baseline's random effects submodel}

\item{q0}{a vector of number of random effects on each initial latent process level}

\item{cholesky}{logical indicating if the variance covariance matrix is parameterized using the cholesky (TRUE, by default) or the correlation (FALSE)}

\item{data_surv}{matrix of Tentry, Event, StatusEvent and covariates for survival models}

\item{data_surv_intY}{matrix of covariates in interaction with Y in survival models}

\item{nYsurv}{number of covariates in interaction with Y in survival models for all nEvents}

\item{basehaz}{baseline hasard function type}

\item{knots_surv}{knots for splines modelling the baseline hazard function}

\item{np_surv}{number of parameters in the survival sub-models !! change if nE>1 !!}

\item{survival}{boolean indicating if survival analysis}

\item{assoc}{type of association between outcomes and times-to-events}

\item{truncation}{boolean indicating if left trucation}

\item{nE}{number of events}

\item{Xsurv1}{design matrix for first event}

\item{Xsurv2}{design matrix for second event}

\item{if_link}{indicates if non linear link is used to transform an outcome}

\item{zitr}{min and max of ordinal outcomes}

\item{ide}{vector of observed values for ordinal outcomes}

\item{tau}{a vector of integers indicating times (including maximum time)}

\item{tau_is}{a vector of integers indicating times for individuals}

\item{modA_mat}{model.matrix for elements of the transistion matrix}

\item{DeltaT}{double that indicates the discretization step}

\item{modA_mat_predGK_t}{design matrix for computing predictions of Y on [0;ti] in Gauss Konrod for all subject}

\item{modA_mat_predGK_t0}{design matrix for computing predictions of Y on [0;t0i] in Gauss Konrod for all subject}

\item{pt_GK_t}{Gauss-Konrod nodes for integration on [0;ti] for all subject}

\item{pt_GK_t0}{Gauss-Konrod nodes for integration on [0;t0i] for all subject}
}
\value{
double
}
\description{
Function that computes the log-likelihood of the observed data
}
