% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matNui_ui}
\alias{matNui_ui}
\title{Function that constructs the matrix lambda_t_j, the value of the processes at time t_j, given the random effects}
\usage{
matNui_ui(
  nD,
  tau_i,
  DeltaT,
  x0i,
  alpha_mu0,
  xi,
  alpha_mu,
  G_mat_A_0_to_tau_i,
  randomeffects,
  zi,
  ordered
)
}
\arguments{
\item{nD}{an integer indicating the number of processes}

\item{tau_i}{vector of integers indicating times}

\item{DeltaT}{double that indicates the discretization step}

\item{x0i}{model.matrix for baseline's submodel}

\item{alpha_mu0}{a vector of parameters associated to the model.matrix for the baseline's submodel (beta)}

\item{xi}{model.matrix for change's submodel}

\item{alpha_mu}{a vector of parameters associated to the model.matrix for the change's submodel (gamma)}

\item{G_mat_A_0_to_tau_i}{matrix containing  Prod(A_t)t=0,tau_i where A_t is the transition
matrix containing at time t}

\item{randomeffects}{vector of individual random effects}

\item{zi}{model.matrix for random change's submodel}

\item{ordered}{indicator if tau_i is ordered or not (usually ordered, except for GK nodes)}
}
\value{
a vector
}
\description{
Function that constructs the matrix lambda_t_j, the value of the processes at time t_j, given the random effects
}
