% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CInLPN2.R
\name{predict.CInLPN2}
\alias{predict.CInLPN2}
\title{Marginal and subject-specific predictions for CInLPN2 objects}
\usage{
\method{predict}{CInLPN2}(object, newdata, MCnr = 10, ...)
}
\arguments{
\item{object}{CInLPN2 object}

\item{newdata}{dataset}

\item{MCnr}{an integer that gives the number of Monte Carlo iterations}

\item{\dots}{optional parameters}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Marginal and subject-specific predictions for CInLPN2 objects
}
