% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.DynNet.R
\name{predict.DynNet}
\alias{predict.DynNet}
\title{Marginal and subject-specific predictions for DynNet objects}
\usage{
\method{predict}{DynNet}(
  object,
  newdata,
  TimeDiscretization = TRUE,
  MCnr = 10,
  ui_hat = NULL,
  ...
)
}
\arguments{
\item{object}{DynNet object}

\item{newdata}{dataset}

\item{TimeDiscretization}{a boolean indicating if the initial time has to be discretized (TRUE by default). When setting to FALSE, it allows to avoid discretization when running univariate model during parameter initialization.}

\item{MCnr}{an integer that gives the number of Monte Carlo iterations}

\item{ui_hat}{matrix of bayesian estimates of random effects}

\item{\dots}{optional parameters}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Marginal and subject-specific predictions for DynNet objects
}
