% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.DynNet.R
\name{predict0.DynNet}
\alias{predict0.DynNet}
\title{Marginal predictions for CInLPN objects}
\usage{
predict0.DynNet(object, newdata, TimeDiscretization = TRUE, MCnr = 10, ...)
}
\arguments{
\item{object}{CInLPN object}

\item{newdata}{dataset}

\item{TimeDiscretization}{a boolean indicating if the inital time have to be discretized. When setting to FALSE, It allows to avoid discretization when running univarite model during parameter initialization.}

\item{MCnr}{an integer that gives the number of Monte Carlo iterations}

\item{\dots}{optional parameters}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Marginal predictions for CInLPN objects
}
