% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.DynNet.R
\name{predict2.DynNet}
\alias{predict2.DynNet}
\title{Marginal and subject-specific predictions for DynNet objects}
\usage{
predict2.DynNet(object, newdata, MCnr = 10, ...)
}
\arguments{
\item{object}{DynNet object}

\item{newdata}{dataset}

\item{MCnr}{an integer that gives the number of Monte Carlo iterations}

\item{\dots}{optional parameters}
}
\value{
list of marginal and subject-specific predictions
}
\description{
Marginal and subject-specific predictions for DynNet objects
}
