% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CInLPN2.R
\name{print.CInLPN2}
\alias{print.CInLPN2}
\title{Print CInLPN2 object}
\usage{
\method{print}{CInLPN2}(x, ...)
}
\arguments{
\item{x}{an CInLPN2 object}

\item{\dots}{optional parameters}
}
\value{
0
}
\description{
Print CInLPN2 object
}
