% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.DynNet.R
\name{print.DynNet}
\alias{print.DynNet}
\title{Print DynNet object}
\usage{
\method{print}{DynNet}(x, ...)
}
\arguments{
\item{x}{an DynNet object}

\item{\dots}{optional parameters}
}
\value{
0
}
\description{
Print DynNet object
}
