% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CInLPN2.R
\name{summary.CInLPN2}
\alias{summary.CInLPN2}
\title{Summary of CInLPN2 object}
\usage{
\method{summary}{CInLPN2}(object, ...)
}
\arguments{
\item{object}{an CInLPN2 object}

\item{\dots}{optional parameters}
}
\value{
0
}
\description{
Summary of CInLPN2 object
}
