% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DynNet.R
\name{summary.DynNet}
\alias{summary.DynNet}
\title{Summary of DynNet object}
\usage{
\method{summary}{DynNet}(object, ...)
}
\arguments{
\item{object}{an DynNet object}

\item{\dots}{optional parameters}
}
\value{
0
}
\description{
Summary of DynNet object
}
