% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeDiscretization.R
\name{TimeDiscretization}
\alias{TimeDiscretization}
\title{function that discretized a vector of time with a given delta.}
\usage{
TimeDiscretization(
  rdata,
  subject,
  fixed_X0.models,
  randoms_X0.models,
  fixed_DeltaX.models,
  randoms_DeltaX.models,
  mod_trans.model,
  outcomes,
  predictors = NULL,
  Time,
  Delta
)
}
\arguments{
\item{rdata}{input data which time point has to be discretized}

\item{subject}{subject identifiant}

\item{fixed_X0.models}{specification of fixed effects for baseline level in longitudinal models}

\item{randoms_X0.models}{specification of random effects for baseline level in longitudinal models}

\item{fixed_DeltaX.models}{specification of fixed effects for slope in longitudinal models}

\item{randoms_DeltaX.models}{specification of random effects for slope in longitudinal models}

\item{mod_trans.model}{specification of transition matrix between latent processes}

\item{outcomes}{a vector of outcomes names}

\item{predictors}{independent variables  to be be included in the modeling}

\item{Time}{colname indicating the time}

\item{Delta}{discretized time step}
}
\value{
a discretized  time vector
}
\description{
function that discretized a vector of time with a given delta.
}
