% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynNet.estim.R
\name{DynNet.estim}
\alias{DynNet.estim}
\title{Title}
\usage{
DynNet.estim(
  K,
  nD,
  mapping.to.LP,
  data,
  if_link = if_link,
  cholesky = FALSE,
  DeltaT = 1,
  MCnr = NULL,
  MCnr2 = NULL,
  nmes = NULL,
  data_surv = NULL,
  paras,
  maxiter = 500,
  nproc = 1,
  epsa = 1e-04,
  epsb = 1e-04,
  epsd = 0.001,
  print.info = FALSE
)
}
\arguments{
\item{K}{number of outcomes}

\item{nD}{number of latent processes}

\item{mapping.to.LP}{indicates which outcome measured which latent process, it is a mapping table between
outcomes and latents processes}

\item{data}{indicates the data frame containing all the variables for estimating the model}

\item{if_link}{indicates if non linear link is used to transform an outcome}

\item{DeltaT}{indicates the discretization step}

\item{paras}{initial values for parameters}

\item{maxiter}{maximum iteration}

\item{nproc}{number of processor to be used for running this package, default value is 1}

\item{epsa}{threshold for the convergence criterion on the parameters, default value is 1.e-4}

\item{epsb}{threshold for the convergence criterion on the likelihood, default value is 1.e-4}

\item{epsd}{threshold for the convergence criterion on the derivatives, default value is 1.e-3}

\item{print.info}{to print information during the liklihood optimization, default value is FALSE}
}
\value{
DynNet object
}
\description{
Title
}
